진원이 버전과 통합
