+++ backend/app.js
@@ -3,7 +3,7 @@ const express = require("express");
 const session = require("express-session");
 const bodyParser = require("body-parser");
 const cookieParser = require("cookie-parser");
-
+const cors = require("cors");
 
 // Create the Express app
 const app = express();
@@ -31,22 +31,22 @@ app.use(
 );
 
 //  api 함수 등록
-const loginRouter = require("./DongUk/routes/login");
-const logoutRouter = require("./DongUk/routes/logout");
-const registerRouter = require("./DongUk/routes/register");
-const searchRouter = require("./DongUk/routes/search"); // Import the search router
-const nowMoviesRouter = require("./DongUk/routes/nowPlayMovie"); // movies.js 파일에서 라우터 가져오기
-const popularMoviesRouter = require("./DongUk/routes/popularMovie"); // popularMovie.js 파일에서 라우터 가져오기
-const isDarkRouter = require("./DongUk/routes/isDark"); // isDark.js 파일에서 라우터 가져오기
-const setDarkRouter = require("./DongUk/routes/setDark"); // isDark.js 파일에서 라우터 가져오기
-const detailMovieRouter = require("./DongUk/routes/detailMovie"); // detailMovie.js 파일에서 라우터 가져오기
-const favoritesRouter = require("./DongUk/routes/addfavorites"); // favorites.js 파일에서 라우터 가져오기
-const removeFavoritesRouter = require("./DongUk/routes/removefavorites"); // favorites.js 파일에서 라우터 가져오기
-const showFavoritesRouter = require("./DongUk/routes/showfavorites"); // favorites.js 파일에서 라우터 가져오기
-const reviewRouter = require("./DongUk/routes/review"); // review.js 파일에서 라우터 가져오기
-const QnARouter = require("./DongUk/routes/questions"); // QnA.js 파일에서 라우터 가져오기
-const personalRoute = require('./routes/personal'); // 회원정보 받아오기
-const personalChangeRoute = require('./routes/personalChange'); // 회원정보 수정하기
+const loginRouter = require("./routes/login");
+const logoutRouter = require("./routes/logout");
+const registerRouter = require("./routes/register");
+const searchRouter = require("./routes/search"); // Import the search router
+const nowMoviesRouter = require("./routes/nowPlayMovie"); // movies.js 파일에서 라우터 가져오기
+const popularMoviesRouter = require("./routes/popularMovie"); // popularMovie.js 파일에서 라우터 가져오기
+const isDarkRouter = require("./routes/isDark"); // isDark.js 파일에서 라우터 가져오기
+const setDarkRouter = require("./routes/setDark"); // isDark.js 파일에서 라우터 가져오기
+const detailMovieRouter = require("./routes/detailMovie"); // detailMovie.js 파일에서 라우터 가져오기
+const favoritesRouter = require("./routes/addfavorites"); // favorites.js 파일에서 라우터 가져오기
+const removeFavoritesRouter = require("./routes/removefavorites"); // favorites.js 파일에서 라우터 가져오기
+const showFavoritesRouter = require("./routes/showfavorites"); // favorites.js 파일에서 라우터 가져오기
+const reviewRouter = require("./routes/review"); // review.js 파일에서 라우터 가져오기
+const QnARouter = require("./routes/questions"); // QnA.js 파일에서 라우터 가져오기
+const personalRoute = require("./routes/personal"); // 회원정보 받아오기
+const personalChangeRoute = require("./routes/personalChange"); // 회원정보 수정하기
 
 //회원가입 로그인 로그아웃 관련 api
 app.use("/auth/logout", logoutRouter);
@@ -69,8 +69,8 @@ app.use("/", reviewRouter); // '/routes/review' 경로에 라우터 적용
 //QnA 관련 api
 app.use("/", QnARouter); // '/routes/question' 경로에 라우터 적용
 // '/routes/personal' 경로에 라우터 적용 -> 회원정보 불러오기
-app.use('/users', personalRoute); 
-app.use('/users', personalChangeRoute);
+app.use("/users", personalRoute);
+app.use("/users", personalChangeRoute);
 
 app.get("/", (req, res) => {
   if (req.session.num === undefined)
@@ -79,7 +79,6 @@ app.get("/", (req, res) => {
   else req.session.num += 1;
   res.send(JSON.stringify(req.session));
 });
-
 
 // Start the server
 const port = 3000;
