+++ backend/data/user.json
@@ -1,8 +1,8 @@
 [
   {
     "username": "apple",
-    "password": "password3",
-    "nickname": "James",
+    "password": "password4",
+    "nickname": "bubble",
     "userID": "user3",
     "isDark": false
   },
