+++ backend/frontend/babel.config.js
