+++ backend/frontend/public/favicon.ico
