+++ backend/frontend/src/assets/logo.png
