+++ backend/helpers/cookieHelper.js
