+++ backend/package.json
@@ -5,6 +5,7 @@
   "main": "app.js",
   "dependencies": {
     "axios": "^1.4.0",
+    "body-parser": "^1.20.2",
     "cookie-parser": "^1.4.6",
     "cors": "^2.8.5",
     "dotenv": "^16.3.1",
@@ -13,17 +14,18 @@
     "express-session": "^1.17.3",
     "memorystore": "^1.6.7",
     "mysql2": "^3.5.2",
-    "nodemon": "^3.0.1",
     "pm2": "^5.3.0",
     "request": "^2.88.2",
     "xlsx": "^0.18.5"
   },
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "start": "nodemon app"
   },
   "author": "",
   "license": "ISC",
   "devDependencies": {
-    "morgan": "^1.10.0"
+    "morgan": "^1.10.0",
+    "nodemon": "^3.0.1"
   }
 }
