+++ backend/routes/addfavorites.js
