+++ backend/routes/logout.js
