+++ backend/routes/nowPlayMovie.js
