+++ backend/routes/removefavorites.js
