+++ backend/routes/showfavorites.js
